Drop two no-op (u8) casts on getObjGroupStatus calls
MapEventInterface.getObjGroupStatus is declared u8 (mapEventTypes.h:18),
so the call already delivers a u8 in r3 and the cast is noise at the two
sites that feed the result straight into an if-condition.

The third site found by the same sweep, 294.c:813, is LOAD-BEARING and
was reverted: there the result is assigned into a u8 local
(groupStatus), and the explicit cast makes MWCC emit the truncation that
the plain assignment does not. Casts landing in a local stay.

Both objects byte-identical (objcopy -R .comment, then
objdump -M gekko -s -t -r); ninja all_source and ninja clean, main.dol OK.

+++ src/dlls/objects/553_DFP_LevelCo/DFP_LevelCo.c
@@ -96,7 +96,7 @@ void DFP_LevelControl_updateMapAct2(GameObject* obj)
     }
     if (mainGetBit(0x7a1) != 0)
     {
-        if ((u8)(*gMapEventInterface)->getObjGroupStatus((obj)->anim.mapEventSlot, 6) == 0)
+        if ((*gMapEventInterface)->getObjGroupStatus((obj)->anim.mapEventSlot, 6) == 0)
         {
             (*gMapEventInterface)->setObjGroupStatus((obj)->anim.mapEventSlot, 6, 1);
         }
